+++ i18n/sample/src/ZoweUSSNode.i18n.json
@@ -2,5 +2,9 @@
     "getChildren.error.invalidNode": "Invalid node",
     "getChildren.error.response": "Retrieving response from ",
     "getChildren.responses.error.response": "The response from Zowe CLI was not successful",
-    "getChildren.responses.open": "Open"
+    "getChildren.responses.open": "Open",
+    "deleteUSSNode.quickPickOption": "Are you sure you want to delete ",
+    "deleteUSSNode.showQuickPick.yes": "Yes",
+    "deleteUSSNode.showQuickPick.no": "No",
+    "deleteUSSNode.error.node": "Unable to delete node: "
 }
